1.8.10

+++ src/main/java/cn/superiormc/ultimateshop/objects/items/products/ObjectSingleProduct.java
@@ -67,7 +67,7 @@ public class ObjectSingleProduct extends AbstractSingleThing {
                     "sell-times-server",
                     String.valueOf(serverSellTimes));
         }
-        double cost = MathUtil.doCalculate(TextUtil.withPAPI(tempVal1, player));
+        double cost = MathUtil.doCalculate(TextUtil.withPAPI(tempVal1, player)).doubleValue();
         if (singleSection.getString("max-amount") != null) {
             double maxAmount = Double.parseDouble(TextUtil.withPAPI(singleSection.getString("max-amount"), player));
             if (cost > maxAmount) {
